+++ src/onegov.org/onegov/org/forms/ticket.py
@@ -4,7 +4,7 @@ from onegov.form.fields import UploadFileWithORMSupport
 from onegov.form.filters import strip_whitespace
 from onegov.form.validators import FileSizeLimit
 from onegov.org import _
-from wtforms import TextAreaField
+from wtforms import BooleanField, TextAreaField
 from wtforms import validators
 
 
@@ -24,3 +24,18 @@ class TicketNoteForm(Form):
             validators.Optional(),
             FileSizeLimit(10 * 1000 * 1000)
         ])
+
+
+class TicketChatMessageForm(Form):
+
+    text = TextAreaField(
+        label=_("Message"),
+        description=_("Your message"),
+        validators=[validators.InputRequired()],
+        filters=(strip_whitespace, ),
+        render_kw={'rows': 5})
+
+
+class InternalTicketChatMessageForm(TicketChatMessageForm):
+
+    notify = BooleanField(label=_("Notify me about replies"))