+++ src/onegov.org/onegov/org/tests/test_views.py
@@ -141,7 +141,7 @@ def test_login(client):
     login_page.form['password'] = 'hunter2'
 
     index_page = login_page.form.submit().follow()
-    assert "Sie wurden eingeloggt" in index_page.text
+    assert "Sie wurden angemeldet" in index_page.text
 
     links = index_page.pyquery('.globals a.logout')
     assert links.text() == 'Logout'
@@ -207,7 +207,7 @@ def test_reset_password(client):
 
     login_page.form['username'] = 'admin@example.org'
     login_page.form['password'] = 'new_password'
-    assert "Sie wurden eingeloggt" in login_page.form.submit().follow().text
+    assert "Sie wurden angemeldet" in login_page.form.submit().follow().text
 
 
 def test_unauthorized(client):
@@ -950,8 +950,8 @@ def test_tickets(client):
     message = message.get_payload(0).get_payload(decode=True)
     message = message.decode('iso-8859-1')
 
-    assert 'FRM-' in message
-    assert '/status' not in message
+    assert "Ihr Ticket wurde geschlossen" in message
+    assert '/status' in message
 
     assert 'FRM-' in status_page
     assert 'Offen' in status_page
@@ -975,7 +975,7 @@ def test_tickets(client):
     message = message.get_payload(0).get_payload(decode=True)
     message = message.decode('iso-8859-1')
 
-    assert 'FRM-' in message
+    assert "Ihr Ticket wurde wieder ge" in message
     assert '/status' in message
 
 
@@ -2505,7 +2505,6 @@ def test_submit_event(client):
         assert (start_date + timedelta(days=days)).strftime('%d.%m.%Y') in \
             message
     assert "Ihre Veranstaltung wurde angenommen" in message
-    assert ticket_nr in message
 
     # Close ticket
     ticket_page.click("Ticket abschliessen").follow()
@@ -2515,7 +2514,8 @@ def test_submit_event(client):
     assert message.get('to') == "test@example.org"
     message = message.get_payload(0).get_payload(decode=True)
     message = message.decode('utf-8')
-    assert ticket_nr in message
+
+    assert "Ihr Ticket wurde geschlossen" in message
 
 
 def test_edit_event(client):
@@ -3212,7 +3212,7 @@ def test_registration(client):
     assert "Konto wurde bereits aktiviert" in client.get(url).follow()
 
     logged_in = client.login('user@example.org', 'p@ssw0rd').follow()
-    assert "eingeloggt" in logged_in
+    assert "angemeldet" in logged_in
 
 
 def test_registration_disabled(client):
@@ -4841,6 +4841,9 @@ def test_ticket_notes(client):
 
     assert "Looks like example input to me" in page
 
+    # make sure we have a badge with the number of notes
+    assert page.pyquery('.counts-note .counts-value').text() == '1'
+
     # edit the note
     note_url_ex = re.compile(r'http://localhost/ticket-notes/[A-Z0-9]+')
     note_url = note_url_ex.search(str(page)).group()
@@ -4884,3 +4887,209 @@ def test_ticket_notes(client):
     page = client.get('/timeline')
     assert "The profile is actually okay" in page
     assert "Test.txt" in page
+
+
+def test_ticket_chat(client):
+    collection = FormCollection(client.app.session())
+    collection.definitions.add('Profile', definition=textwrap.dedent("""
+        First name * = ___
+        Last name * = ___
+        E-Mail * = @@@
+    """), type='custom')
+
+    transaction.commit()
+
+    # submit a form
+    client.login_admin()
+
+    page = client.get('/forms').click('Profile')
+    page.form['first_name'] = 'Foo'
+    page.form['last_name'] = 'Bar'
+    page.form['e_mail'] = 'foo@bar.baz'
+    page = page.form.submit().follow().form.submit().follow()
+
+    # make sure a ticket has been created
+    assert 'FRM-' in page
+    assert 'ticket-state-open' in page
+
+    # to extract the messages from the page
+    def extract_messages(page):
+        text = page.pyquery('[data-feed-data]').attr('data-feed-data')
+        data = json.loads(text)
+
+        return data['messages']
+
+    # we should see the initial ticket state in the messages
+    msgs = extract_messages(page)
+    assert len(msgs) == 1
+    assert "Ticket eröffnet" in msgs[0]['html']
+
+    # we should also see one e-mail at this point
+    assert len(client.app.smtp.outbox) == 1
+
+    # at this point we have the option to send an initial message
+    page.form['text'] = "I spelt my name wrong: it's Baz, not Bar"
+    page = page.form.submit().follow()
+
+    # which should show up in the status page and the ticket page
+    status_url = page.request.url
+
+    msgs = extract_messages(page)
+    assert len(msgs) == 2
+    assert "I spelt my name wrong" in msgs[1]['html']
+
+    # we should see the same infos on the ticket page
+    page = client.get('/tickets/ALL/open').click("Annehmen").follow()
+
+    msgs = extract_messages(page)
+    assert len(msgs) == 3
+    assert "I spelt my name wrong" in msgs[1]['html']
+    assert "Ticket angenommen" in msgs[2]['html']
+
+    assert page.pyquery('.counts-external .counts-value').text() == '1'
+
+    # no e-mail will have been created for this message, as there's no
+    # recipient we could send it to
+    assert len(client.app.smtp.outbox) == 1
+
+    # add a note and let's ensure that the status page does not contain it
+    page = page.click("Neue Notiz")
+    page.form['text'] = "Contacting the user"
+    page = page.form.submit().follow()
+
+    msgs = extract_messages(page)
+    assert len(msgs) == 4
+    assert "Contacting the user" in msgs[3]['html']
+
+    msgs = extract_messages(client.get(status_url))
+    assert len(msgs) == 3
+    assert "Contacting the user" not in json.dumps(msgs)
+
+    # answer the user, enable notifications to the editor
+    ticket_url = page.request.url
+
+    page = page.click("Nachricht senden")
+    page.form['text'] = "I will correct your name"
+    page.form.get('notify').checked = True
+    page.form.submit()
+
+    # ensure the user sees the message on the status page
+    msgs = extract_messages(client.get(status_url))
+
+    assert len(msgs) == 4
+    assert "I will correct your name" in msgs[-1]['html']
+
+    # ensure the user sees the message in an e-mail
+    assert len(client.app.smtp.outbox) == 2
+    assert "I will correct your name" in client.get_email(1)
+
+    # make sure the reply-to story is good
+    mail = client.app.smtp.outbox[1]
+    assert mail['To'] == 'foo@bar.baz'
+    assert mail['From'] == 'Govikon <mails@govikon.ch>'
+    assert mail['Reply-To'] == 'Govikon <admin@example.org>'
+
+    # ensure the editor sees the messages in the ticket view
+    msgs = extract_messages(client.get(ticket_url))
+
+    assert len(msgs) == 5
+    assert "I will correct your name" in msgs[-1]['html']
+
+    page = client.get(ticket_url)
+    assert page.pyquery('.counts-internal .counts-value').text() == '1'
+
+    # answer the editor
+    page = client.get(status_url)
+    page.form['text'] = 'Great, thanks!'
+    page.form.submit()
+
+    # the editor should get a notification with the last message
+    assert len(client.app.smtp.outbox) == 3
+
+    mail = client.get_email(2)
+    assert "I will correct your name" in mail
+    assert "Great, thanks!" in mail
+
+    # send a final message from the editor, with no notification
+    page = client.get(ticket_url).click("Nachricht senden")
+    page.form['text'] = "You're welcome"
+    page.form.get('notify').checked = False
+    page.form.submit()
+
+    # the user always gets a notification
+    assert len(client.app.smtp.outbox) == 4
+
+    # but now, the answering wont create one
+    page = client.get(status_url)
+    page.form['text'] = 'Can I ask you something else?'
+    page.form.submit()
+
+    assert len(client.app.smtp.outbox) == 4
+
+    # though it will still show up in the list
+    msgs = extract_messages(client.get(ticket_url))
+
+    assert len(msgs) == 8
+    assert "Can I ask you something else?" in msgs[-1]['html']
+
+    # close the ticket
+    status_before = client.get(status_url)
+    assert "Ticket wurde bereits geschlossen" not in status_before
+
+    page = client.get(ticket_url)
+    page.click("Ticket abschliessen")
+
+    # we should no longer be able to send messages
+    page = client.get(status_url)
+    len(page.forms) == 1
+
+    status_before.form['text'] = 'Foo'
+    page = status_before.form.submit()
+
+    assert "Ticket wurde bereits geschlossen" in page
+
+    # verify the same for the editor
+    page = client.get(ticket_url)
+    assert "Zu geschlossenen Tickets können keine Nachrichten" in page
+
+    page = client.get(ticket_url + '/message-to-submitter')
+    page.form['text'] = 'One more thing'
+    page = page.form.submit()
+
+    assert "Ticket wurde bereits geschlossen" in page
+
+
+def test_ticket_chat_search(client_with_es):
+    client = client_with_es
+
+    collection = FormCollection(client.app.session())
+    collection.definitions.add('Profile', definition=textwrap.dedent("""
+        First name * = ___
+        Last name * = ___
+        E-Mail * = @@@
+    """), type='custom')
+
+    transaction.commit()
+
+    # submit a form
+    client.login_admin()
+
+    page = client.get('/forms').click('Profile')
+    page.form['first_name'] = 'Foo'
+    page.form['last_name'] = 'Bar'
+    page.form['e_mail'] = 'foo@bar.baz'
+    page = page.form.submit().follow().form.submit().follow()
+
+    # send a message that should be findable through the search
+    page.form['text'] = "I spelt my name wrong: it's deadbeef"
+    page = page.form.submit().follow()
+
+    # at this point logged in users should find the ticket by 'deadbeef'
+    client.app.es_client.indices.refresh(index='_all')
+
+    page = client.get('/search?q=deadbeef')
+    assert 'Foo' in page
+
+    # but anonymous users should not
+    page = client.spawn().get('/search?q=deadbeef')
+    assert 'Foo' not in page