+++ src/onegov.org/onegov/org/views/message.py
@@ -86,6 +86,7 @@ def view_messages_feed(self, request):
             {
                 'id': m.id,
                 'type': m.type,
+                'subtype': m.subtype or '',
                 'date': ', '.join((
                     layout.format_date(m.created, 'weekday_long'),
                     layout.format_date(m.created, 'date_long')